requirements-dev.txt
pylint
mypy
yapf
flake8
types-PyYAML
pandas-stubs
yamllint
pyright
